move links group to customized links
